Added error checking to the scanner
Scanner no longer returns a vector of tokens. Instead it accepts a
pointer to a vector as input and returns an int as a result code.
Scanner will now throw an error if an invalid token is generated.
Added dfa to generate tokens for whitespace and control characters.
Added cleanup to free memory on program termination.
Fixed a bug that was causing the scanner to initialize with 0 dfas.