Initial demonstration of AST structure
I've added classes to represent some of the non-terminals in the AST and
I've added an ast class to handle constructing the AST.

This is the basic strategy that I was thinking we could follow to create
the entire AST.

It would make it really easy to implement synthesized attributes as each
node in the tree can call methods implemented by its children to get the
necessary information.

As far as inherited attributes go, the children can implement methods
that will be called by the parents to set the relevant attributes.

We can talk about all the details of how e do that later, but what do
you guys think about the general structure I've started here?

If everyone thinks this is a good way to go, I'll work on churning out a
bunch more AST classes this weekend and we can divide up the parts of
the grammar to work on if other people want to work on AST classes as
well.